fix regression comment
