+++ fe/fe-core/src/main/java/org/apache/doris/nereids/rules/exploration/mv/StructInfo.java
@@ -315,7 +315,7 @@ public class StructInfo {
         // Get mapped table id in relation and set
         BitSet tableBitSet = new BitSet();
         for (CatalogRelation relation : relationList) {
-            tableBitSet.set(cascadesContext.getTableId(relation.getTable()));
+            tableBitSet.set(cascadesContext.getStatementContext().getTableId(relation.getTable()));
         }
         return new StructInfo(originalPlan, originalPlanId, hyperGraph, valid, topPlan, bottomPlan,
                 relationList, relationIdStructInfoNodeMap, null, shuttledHashConjunctsToConjunctsMap,
